# docs/openapi.yaml
openapi: 3.1.0
info:
  title: BidOps API
  version: 0.1.0
  summary: BidOps MVP OpenAPI 1차 초안
  description: >
    공공/민간 입찰·제안 업무를 위한 근거 기반 RFP 분석 및 수주 운영 시스템 API.
    문서 업로드 → 비동기 분석 → 요구사항 검토 → 체크리스트/질의/산출물 관리 흐름을 기준으로 정의한다.
servers:
  - url: /api/v1
    description: API v1
...
security:
  - bearerAuth: []
paths:
...
  /projects/{projectId}:
    get:
      tags: [Projects]
      summary: 프로젝트 상세 조회
      operationId: getProject
      parameters:
        - $ref: '#/components/parameters/ProjectId'
      responses:
        '200':
          description: 프로젝트 상세
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ProjectResponse' }
    patch:
      tags: [Projects]
      summary: 프로젝트 수정
      operationId: updateProject
      parameters:
        - $ref: '#/components/parameters/ProjectId'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ProjectUpdateRequest' }
      responses:
        '200':
          description: 프로젝트 수정 완료
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ProjectResponse' }
...
  /projects/{projectId}/documents:
    get:
      tags: [Documents]
      summary: 문서 목록 조회
      operationId: listDocuments
      parameters:
        - $ref: '#/components/parameters/ProjectId'
        - in: query
          name: type
          schema: { $ref: '#/components/schemas/DocumentType' }
        - in: query
          name: parse_status
          schema: { $ref: '#/components/schemas/DocumentParseStatus' }
      responses:
        '200':
          description: 문서 목록
          content:
            application/json:
              schema: { $ref: '#/components/schemas/DocumentListResponse' }
    post:
      tags: [Documents]
      summary: 문서 업로드
      operationId: uploadDocument
      parameters:
        - $ref: '#/components/parameters/ProjectId'
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file, type]
              properties:
                file: { type: string, format: binary }
                type: { $ref: '#/components/schemas/DocumentType' }
                version_note: { type: string }
      responses:
        '201':
          description: 문서 업로드 완료
          content:
            application/json:
              schema: { $ref: '#/components/schemas/DocumentResponse' }
  /projects/{projectId}/documents/{documentId}:
    get:
      tags: [Documents]
      summary: 문서 상세 조회
      operationId: getDocument
      parameters:
        - $ref: '#/components/parameters/ProjectId'
        - $ref: '#/components/parameters/DocumentId'
      responses:
        '200':
          description: 문서 상세
          content:
            application/json:
              schema: { $ref: '#/components/schemas/DocumentResponse' }
    delete:
      tags: [Documents]
      summary: 문서 삭제
      operationId: deleteDocument
      parameters:
        - $ref: '#/components/parameters/ProjectId'
        - $ref: '#/components/parameters/DocumentId'
      responses:
        '200':
          description: 문서 삭제 완료
          content:
            application/json:
              schema: { $ref: '#/components/schemas/SuccessResponse' }
  /projects/{projectId}/documents/{documentId}/versions:
    get:
      tags: [Documents]
      summary: 문서 버전 목록 조회
      operationId: listDocumentVersions
      parameters:
        - $ref: '#/components/parameters/ProjectId'
        - $ref: '#/components/parameters/DocumentId'
      responses:
        '200':
          description: 문서 버전 목록
          content:
            application/json:
              schema: { $ref: '#/components/schemas/DocumentVersionListResponse' }
...
  /projects/{projectId}/audit-logs:
    get:
      tags: [AuditLogs]
      summary: 감사로그 목록 조회
      operationId: listAuditLogs
      parameters:
        - $ref: '#/components/parameters/ProjectId'
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/Size'
        - in: query
          name: actor_user_id
          schema: { type: string }
        - in: query
          name: action_type
          schema: { type: string }
        - in: query
          name: target_type
          schema: { type: string }
        - in: query
          name: date_from
          schema: { type: string, format: date-time }
        - in: query
          name: date_to
          schema: { type: string, format: date-time }
      responses:
        '200':
          description: 감사로그 목록
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AuditLogListResponse' }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    ProjectId:
      in: path
      name: projectId
      required: true
      schema: { type: string }
    DocumentId:
      in: path
      name: documentId
      required: true
      schema: { type: string }
...
    Page:
      in: query
      name: page
      schema: { type: integer, minimum: 1, default: 1 }
    Size:
      in: query
      name: size
      schema: { type: integer, minimum: 1, maximum: 100, default: 20 }
  schemas:
    Meta:
      type: object
      properties:
        page: { type: integer }
        size: { type: integer }
        total_count: { type: integer }
      additionalProperties: true
...
    SuccessResponse:
      type: object
      required: [success, data, meta, error]
      properties:
        success: { type: boolean, const: true }
        data: { type: object, additionalProperties: true }
        meta: { $ref: '#/components/schemas/Meta' }
        error: { type: 'null' }
    ProjectStatus:
      type: string
      enum: [DRAFT, READY, ANALYZING, REVIEWING, SUBMISSION_PREP, CLOSED]
    DocumentParseStatus:
      type: string
      enum: [UPLOADED, PARSING, PARSED, FAILED]
...
    DocumentType:
      type: string
      enum: [RFP, ANNEX, FORM, QNA, PROPOSAL_REFERENCE]
...
    Project:
      type: object
      required: [id, name, client_name, business_name, status, created_by, created_at, updated_at]
      properties:
        id: { type: string }
        name: { type: string }
        client_name: { type: string }
        business_name: { type: string }
        status: { $ref: '#/components/schemas/ProjectStatus' }
        description: { type: string }
        created_by: { type: string }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
    Document:
      type: object
      required: [id, project_id, type, file_name, storage_path, version, parse_status, uploaded_by, uploaded_at]
      properties:
        id: { type: string }
        project_id: { type: string }
        type: { $ref: '#/components/schemas/DocumentType' }
        file_name: { type: string }
        storage_path: { type: string }
        version: { type: integer }
        version_note: { type: string }
        parse_status: { $ref: '#/components/schemas/DocumentParseStatus' }
        page_count: { type: integer, nullable: true, description: "PDF 전체 페이지 수 (파싱 완료 후 채워짐)" }
        uploaded_by: { type: string }
        uploaded_at: { type: string, format: date-time }
        viewer_url: { type: string, description: "PDF 열기/다운로드 URL" }
...
    AuditLog:
      type: object
      required: [id, project_id, actor_user_id, action_type, target_type, target_id, created_at]
      properties:
        id: { type: string }
        project_id: { type: string }
        actor_user_id: { type: string }
        action_type: { type: string }
        target_type: { type: string }
        target_id: { type: string }
        detail_json:
          type: object
          additionalProperties: true
        created_at: { type: string, format: date-time }
...
    ProjectUpdateRequest:
      type: object
      properties:
        name: { type: string }
        client_name: { type: string }
        business_name: { type: string }
        description: { type: string }
...
    ProjectResponse:
      type: object
      required: [success, data, meta, error]
      properties:
        success: { type: boolean, const: true }
        data: { $ref: '#/components/schemas/Project' }
        meta: { $ref: '#/components/schemas/Meta' }
        error: { type: 'null' }
...
    DocumentResponse:
      type: object
      required: [success, data, meta, error]
      properties:
        success: { type: boolean, const: true }
        data: { $ref: '#/components/schemas/Document' }
        meta: { $ref: '#/components/schemas/Meta' }
        error: { type: 'null' }
    DocumentListResponse:
      type: object
      required: [success, data, meta, error]
      properties:
        success: { type: boolean, const: true }
        data:
          type: object
          required: [items, total_count]
          properties:
            items:
              type: array
              items: { $ref: '#/components/schemas/Document' }
            total_count: { type: integer }
        meta: { $ref: '#/components/schemas/Meta' }
        error: { type: 'null' }
    DocumentVersionListResponse:
      type: object
      required: [success, data, meta, error]
      properties:
        success: { type: boolean, const: true }
        data:
          type: object
          required: [items]
          properties:
            items:
              type: array
              items: { $ref: '#/components/schemas/Document' }
        meta: { $ref: '#/components/schemas/Meta' }
        error: { type: 'null' }
...
    AuditLogListResponse:
      type: object
      required: [success, data, meta, error]
      properties:
        success: { type: boolean, const: true }
        data:
          type: object
          required: [items, total_count]
          properties:
            items:
              type: array
              items: { $ref: '#/components/schemas/AuditLog' }
            total_count: { type: integer }
        meta: { $ref: '#/components/schemas/Meta' }
        error: { type: 'null' }
...